move css to styles folder
