+++ src/App.js
@@ -10,7 +10,7 @@ import { APOD_API } from "./data/api.nasa.gov/api";
 import account from "./data/api.nasa.gov/account";
 
 /// styles ///
-import "./App.css";
+import "./styles/App.css";
 
 /***************************************
   STATES
