+++ src/index.js
@@ -7,7 +7,7 @@ import App from "./App";
 
 /// styles ///
 // import 'bootstrap/dist/css/bootstrap.min.css';
-import "./index.css";
+import "./styles/index.css";
 
 /// do the thing ///
 ReactDOM.render (<App /> , document.getElementById ("root"));
